Convert DebugApi to Typescript
